Add task Join
Add a task for joining the rows of two tables on the values of one or more
columns. Either an inner join or a left outer join can be performed. Matching
rows can be included in the result as is or their columns can be expanded as
columns of the resulting table.