Load: Refactor XML loading types
XML column sources couldn't be defined in Frends as the properties did not
show in the UI. This seems to indicate that Frends only lists the properties
of the class itself, not the properties of base classes. Also Frends doesn't
seem to be able to handle generic types for properties.

Define identical properties in the non-generic XmlColumnSource type and use
the generic XmlColumnSource<TPath> internally.